[#80] task_definition route 레이어 분리
