+++ src/domain/task_definition/routes/mod.rs
@@ -1,0 +1,1 @@
+pub mod http;
